Remove the IndexingCriteria trick
Now each extractor is chosen only based on mimetype. If no mimetype is
found then the search is run again, but this time startsWith is used
instead of an absolute match.

This way we can support exact plugins such as "text/something" and have
a generic plugin which works on all text files by specifying "text/"