Add ExtractionFlags
ExtractionFlags allow us to configure what all information should be
extracted by the ExtractionPlugins. Currently the only supported stuff
is -

* Extract MetaData
* Extract Plain Text

Quite often we do not want to parse all of the plain text and only want
some of the metadata. In this case it makes sense to avoid extracting
the plain text. Specially for pdfs.

Also, added a test for the plain text indexer.