Add an Extractor class
This Extractor class is just a light wrapper over the ExtractorPlugin
class. It however adds a nice level of indirection. Clients now no
longer need to directly deal with plugins and the API is cleaner. This
also allows us the liberty to easily add newer plugin interfaces without
breaking the code of people using this API.