fix build/test by making translation properties testable
propertyinfotest iterates through all properties to make sure beginning
and end are set correctly and that all properties in-between have a name
and description.

this was not the case for new translation properties, making the test fail

due to the nature of this test iterating through everything it would also
be very useful if the compiler warned about missing propertyinfos which
is achieved by not using a default case in the respective switch.

also a note comment was added pointing out that an observant developer
adding a new entry also needs to add it in the fromName function.

REVIEW: 121266