Fix Python dependency in test scripts for KFileMetaData (bug 375472)
Summary:
KFileMetaData test scripts are usable in both deprecated Python 2, and new Python 3.  This removes 'python2' from the shebang and uses the regular 'python' interpreter, thereby removing the requirement of Python 2 being on a system to be able to test KFileMetaData.

Additionally, it removes unused imports.

Test Plan:
This patch was applied to:
* 5.23
* 5.26
* 5.29

It was tested on Linux kernel 4.4 using:
* x86_64
* x86_32
* PowerPC

With the following Python versions:
* 2.7.10
* 3.4.4
* 3.5.1

It was tested on glibc and musl libc.

Reviewers: mgallien

Reviewed By: mgallien

Subscribers: #frameworks

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D4377