use own extractor for testing the taglibwriter
Summary:
Instead of querying the tags directly through taglib,
use the taglibextractor to read the written tags for the tests.
This will become useful once additional tags are added to the
taglibwriter that require manual querying and reading.
As a bonus, the now unneeded custom TString to QString function
is removed.

Test Plan: compiles and tests pass

Reviewers: mgallien, bruns, svuorela

Reviewed By: bruns

Subscribers: kde-frameworks-devel, #baloo

Tags: #frameworks, #baloo

Differential Revision: https://phabricator.kde.org/D16111