[Extractor] Add metadata to extractors
Summary:
This adds extractor metadata in a backwards and forward compatible way.

There are several use cases for this metadata:
- Delayed loading of extractor plugins - currently, all extractors are
  loaded and and initialized when an ExtractorCollection is created.
- Versioning information - e.g. Baloo would benefit from versioning
  information, to reindex affected files after an extractor has been
  updated.

Although it would be possible to extend the extractor plugin interface
with a method for each relevant property, it would require a bump of
the plugin inteface version each time the interface is extended.

CCBUG: 404171
See: T9867, T8079

Test Plan: ctest

Reviewers: #baloo, #frameworks, ngraham, astippich, poboiko

Reviewed By: astippich

Subscribers: kde-frameworks-devel

Tags: #frameworks, #baloo

Differential Revision: https://phabricator.kde.org/D19109