Rewrite the taglib extractor to use the generic PropertyMap interface
Summary:
Rewrite the taglib extractor to use taglib's
PropertyMap. Since this largely unifies the handling of the
different tag formats, but not quite, a lot of code is removed.
The resulting code is also faster. Additionally, this avoids the
usage of a FileRef object, which fixes a potential crash due to
a known bug in taglib.

BUG: 403902

Test Plan: all tests pass

Reviewers: ngraham, bruns, mgallien

Reviewed By: bruns

Subscribers: smithjd, kde-frameworks-devel, #baloo

Tags: #frameworks, #baloo

Differential Revision: https://phabricator.kde.org/D18826