Add a helper to determine actual supported parent mime type
Summary:
The ExtractorCollection::fetchExtractors method falls back to ancestor
mime types in case a type is not supported directly. For most extractors
this is no problem, as these ignore the passed in mime type and rely
on the backing library to figure out the type, while the taglibextractor
requires one of the supported (container) types.

For most cases the input type is returned, but it falls back to
inheritance, matching the logic from ExtractorCollection::fetchExtractors.

Reviewers: #baloo, #frameworks, ngraham, astippich

Reviewed By: astippich

Subscribers: kde-frameworks-devel

Tags: #frameworks, #baloo

Differential Revision: https://phabricator.kde.org/D20138