[UserMetaData] Add method to query which attributes are set
Summary:
Many files have no UserMetaData attributes set at all, or only a subset
of the available ones. Allow to bulk query which attributes are set to
short cut the one-by-one query.

Test Plan:
ctest (on Linux)

Code paths for Mac and FreeBSD are fully implemented, but untested. The
Windows implementation is mostly a stub, but functional conforming with
the documentation.

Reviewers: #baloo, #frameworks, astippich, ngraham

Reviewed By: astippich

Subscribers: arrowd, #freebsd, #windows, kde-frameworks-devel

Tags: #frameworks, #baloo

Differential Revision: https://phabricator.kde.org/D20967