[EPubExtractor] Handle metadata with repeated fields properly
Concatenating the list returned for libepub breaks the value extraction
in several places later:

```
    <dc:date opf:event="publication">2006-03-11</dc:date>
    <dc:date opf:event="conversion">2018-11-01</dc:date>
```
causes a "publication" date which can not be parsed:
> Could not determine correct datetime format from:
> "2012-01-01, conversion: 2014-01-01T01:01:01+00:00"

Only the first "Creator" role is stripped, and only if it is the
main author:
> Author: ill: <Illustrator name>

Add each list entry returned from libepub as a separate entry in the
KFM property map.

Extend the unit test to cover repeated keys in the metadata.