[ExtractionResult] Remove automatic mimetype determination
ExtractionResult is used as a base class extended by e.g.
SimpleExtractionResult in KFM and Baloo::Result and should be a minimal
storage class.

The automatic mime type determination fallback is not mentioned in the
API description in any way; it is not tested; it is not used by either
KFM itself or e.g. Baloo; and it is not failsafe, as the url may be
unreachable for Qt.

Remove it as it is essentially dead code.