Port deprecated QMap methods to QMultiMap
Considering that we have these methods in the public API, we
can not change the return type. Consequently we have to provide an overload
which provides the QMultiMap return value. This overload takes
an enum value to avoid having a ambiguous member function call.

In case one disables deprecated public API, we can privide a default arg.
This way consumers will not have to change their code for KF6.

Task: https://phabricator.kde.org/T14337