Fix loading of external extractors and writers
In 5bbdfc4749f92c510cc82dbd66f5d3b2bc3763ac, the external extractors were mistakenly ported to KPluginMetaData::findPlugins.

However, the logic was broken before, because the dir entries must *not* be plugins, but normal files.
This was introduced in 73da9a53cb500da000c50b164dd5693ef7474041.

Considering that we tell QDir to only list dirs, the check if we have a plugin or not doesn't make sense anyways.

BUG: 463598