Add Krita extractor
This extracts some information from the Krita document such as
canvas width, height and some metadata like the author and
description. The file format is just XML, so it's trivial to parse.

I can't give you a screenshot because it doesn't seem like the file details show up in Plasma 6 yet, but the output of `dump` works:

```
$ ./dump ~/test.kra
/home/josh/test.kra application/x-krita
        KritaExtractor For application/x-krita
                Title: test title (QString)
                Keywords: apples (QString)
                Width: 2177 (int)
                Height: 1428 (int)
                License: CC-BY 4.0 (QString)
                Description: this is a test file, with a test file! (QString)
        -------------
```