Reapply ReplayGain changes
The change was lost in the recent PropertyInfo refactoring, reapply most
of it. The inner double colon is removed, as a label/value like
"ReplayGain: Album Gain: 123" just looks strange, and we also don't do
such a split elsewhere, i.e. we use "GPS latitude: 52'E", not
"GPS: latitute: 52'E".