Fix compilation with latest TagLib git master
The latest taglib `RIFF::WAV::File::tag` now returns `TagLib::Tag *`
instead of `ID3v2::Tag *`. The old behaviour was kept only for backwards
compatibility and was slated to be fixed in TagLib 2.0.

The 1:1 replacement is `RIFF::WAV::File::ID3v2Tag`, which was added in
TagLib 1.9 (released 10 years ago).