[TypeInfo] Improve Type documentation
Better distinguish between file formats which use an encoding
(are "sub-class-of" in shared-mime-info nomenclature) as an
implementation detail; and where the mime-type directly maps to a user
facing `Type` category.

`Text` is actually the outlier here, as it typically is not a Type which
provides any information for the user. Notably, `text/plain` is both
`Text` and `Document`.