[PropertyInfo,TypeInfo] Avoid lookup from id during initilization
The initialization used a lookup from ID during initialization to create
a full PropertyInfo/TypeInfo. As the pointer to the shared private data
suffices to create the PI/TI during the actual lookup, remove the lookup.

This removes the largest part of the initialization cost, and is even
slightly faster during lookup.