plaintextextractor: Read file via QFile
Read file via QFile for both cases: for systems that support O_NOATIME
and for those that do not support this flag. The only difference is the
method of opening QFile, where for the case of O_NOATIME, opening using
the file descriptor `fd` is used.

This change is intended to unify the code for different systems
(Linux C-style and STL C++ style to one Qt-Style via QFile), as well
as simplify it.