Move date parser helper out of ExtractorPlugin, clean it up
The method is only loosely related to the plugins, and there is no
reason for it being an ExtractorPlugin method.

Clean up the repetitive QDateTime::fromString etc. calls and replace
it with a loop over all supported formats.

If the deprecated compat wrapper is removed, the code is also removed
from all plugins not using it.