[TaglibExtractor|Writer] Fix mimetypes
"audio/opus" has never been a valid mime type name in shared-mime-info,
or an alias of a valid mime type. The test only ever worked as the
tests and the extractor both used the same bogus name verbatim.

Likewise for "audio/mpeg3", which has never been a valid name or alias
in shared-mime-info.

"audio/x-mpeg" is a valid alias, but has never been the name of the
type, i.e. there has never been an SMI version which returned
"audio/x-mpeg" instead of the actual "audio/mpeg".