[OfficeExtractor] Remove duplicate findExecutable calls, fix debug output
QStandardPaths::findExecutable is already called for each of the three
helpers in the constructor, no need to repeat the call in `textFromFile`.

This also fixes the bad debug output, if no helper is found, `command`
will be an empty string, and the debug message becomes quite pointless.
Move the debug output to the constructor.