Fix EPubExtractorTest::testRepeated on a non UTC system
Parsing "2012-01-01" from test_repeated.epub results in a QDateTime
with a Qt::LocalTime with "Europe/Berlin" as timezone on my local
machine, which has that as a system timezone.
This throws off EPubExtractorTest::testRepeated, because it compares
with a UTC based time, or in other words:

Actual   (result.properties().value(Property::CreationDate)): QVariant(QDateTime,2012-01-01T00:00:00.000)
Expected (QVariant(dt))                                     : QVariant(QDateTime,2012-01-01T00:00:00.000Z)

This doesn't happen on another system, which runs with a system time
zone set to UTC. Which is probably true for the CI as well and the
reason why it doesn't show there.

To solve this, just use UTC here as well, like the other date format
code paths do. The only problem is, that a ISO 8601 date could contain
a 'Z' for UTC or an offset from UTC. In these case, dateTime's timeSpec
would be Qt::UTC already or Qt::OffsetFromUTC though.