Install dump utility as a debugging aid for users
Users should be able to determine which properties are available
for a file, if the text could be extracted, and how long it took to
extract the contents. The dump utility provides all this.

The utility is installed under the name "kfilemetadata_dump6" to avoid
any conflicts with other binaries.