[autotests] Make usermetadatawritertest idempotent, cleanup test output
The taglibwritertest and usermetadatawritertest were writing to the
source directory, move all output to the binary dir.

Notably the usermetadatawritertest can not use a regular temporary
directory, as this is commonly on tmpfs, which does not have full
extended attrributes support.

This also fixes a bug in the usermetadatawritertest, as that failed
to remove the "metadata-dir", causing failures in subsequent runs for
`QDir::mkdir`.