[PlaintextExtractor] Verify decoded text contains printable characters
In case the decoded text mostly contains control characters and similar,
but hardly any letters, number or punctuation, it is very likely the
file actually contains fairly arbitrary binary data.

This mostly happens when a file starts with a BOM, as it will be
detected as text/plain by the mime database.

BUG: 506570