[PlainTextExtractor] Lower required confidence for UTF-8
KEncodingProber returns a fairly low confidence for UTF-8 texts, even
if it has ruled out any other encoding.

This is noticable mostly on windows, where the fallback
QStringConverter::System differs from UTF-8.

It is also only noticeable for texts outside the ASCII range, as
for ASCII only any Decoder which is a superset works.