[CsvExtractor] Add extractor plugin for handling CSV data
The PlaintextExtractor handles CSV data fairly, but notably fails to
handle numbers correctly, i.e. it can't distinguish between a comma as
field separator and a comma as thousands or decimal separator (as used
depending on the locale).

The unit tests include two CSV exports of an ODS spreadsheet, covering
the most common CSV dialects. More dialects are covered by the CSV helper
methods and corresponding unit test.

The text codec is detected by the same method as used by the plaintext
extractor.