[PlaintextExtractor] Use explicitly sized type for 32bit compatibility
`QIODevice::read(...)` returns `qint64`, while Qt containers like
`QByteArray(View)` use the architecture dependent, possibly smaller
`qsizetype`. This causes narrowing warnings when the size is passed
to the `QByteArrayView` constructor.

As the actual possible return value is upper-bound by the passed input
buffer size, it is safe to explicitly cast it to a smaller signed type
like `int32_t`.

Brace-initilization of the `chunkSize` with the same type provides a
compile time check of the value range.